Updated logos

+++ rasterize.sh
@@ -1,54 +1,64 @@
 #!/bin/sh
 
 # Author:     Héctor Molinero Fernández <hector@molinero.dev>
+# License:    MIT, https://opensource.org/licenses/MIT
 # Repository: https://github.com/hectorm/logo
 
 set -eu
+export LC_ALL='C'
 
-SCRIPT_DIR=$(dirname "$(readlink -f "$0")")
-TARGET_DIR=${SCRIPT_DIR}/bitmap
+SCRIPT_DIR="$(CDPATH='' cd -- "$(dirname -- "${0:?}")" && pwd -P)"
+TARGET_DIR="${SCRIPT_DIR:?}/bitmaps/"
 
 main() {
-	rm -rf "${TARGET_DIR}"
-	mkdir -p "${TARGET_DIR}"
+	rm -rf "${TARGET_DIR:?}"
+	mkdir -p "${TARGET_DIR:?}"
 
-	for i in $(seq 0 "$(jq '.vectors|length-1' "${SCRIPT_DIR}/build.json")"); do
-		name=$(jq       -r --arg i "${i}" '.vectors[$i|tonumber].name' "${SCRIPT_DIR}/build.json")
-		file=$(jq       -r --arg i "${i}" '.vectors[$i|tonumber].file' "${SCRIPT_DIR}/build.json")
-		background=$(jq -r --arg i "${i}" '.vectors[$i|tonumber].background' "${SCRIPT_DIR}/build.json")
-		margin=$(jq     -r --arg i "${i}" '.vectors[$i|tonumber].margin' "${SCRIPT_DIR}/build.json")
-		totalSizes=$(jq -r --arg i "${i}" '.vectors[$i|tonumber].sizes|length-1' "${SCRIPT_DIR}/build.json")
+	vectors="$(cat "${SCRIPT_DIR:?}/rasterize.json")"
+	vectorTotal="$(jq -nr --argjson d "${vectors:?}" '$d|length-1')"
 
-		for j in $(seq 0 "${totalSizes}"); do
-			size=$(jq -r --arg i "${i}" --arg j "${j}" '.vectors[$i|tonumber].sizes[$j|tonumber]' "${SCRIPT_DIR}/build.json")
+	vectorIndex='0'
+	while [ "${vectorIndex:?}" -le "${vectorTotal:?}" ]; do
+		vector="$(jq -nr --argjson d "${vectors:?}" --arg i "${vectorIndex:?}" '$d[$i|tonumber]')"
+		src="$(jq -nr --argjson d "${vector:?}" '$d.src')"
+		name="$(jq -nr --argjson d "${vector:?}" '$d.name')"
+		heightTotal="$(jq -nr --argjson d "${vector:?}" '$d.heights|length-1')"
+		backgroundTotal="$(jq -nr --argjson d "${vector:?}" '$d.backgrounds|length-1')"
+		marginTotal="$(jq -nr --argjson d "${vector:?}" '$d.margins|length-1')"
 
-			inkscape --without-gui \
-				-w "${size}" -h "${size}" \
-				-e "${TARGET_DIR}/${name}_x${size}.png" \
-				"${SCRIPT_DIR}/${file}"
+		backgroundIndex='0'
+		while [ "${backgroundIndex:?}" -le "${backgroundTotal:?}" ]; do
+			background="$(jq -nr --argjson d "${vector:?}" --arg i "${backgroundIndex:?}" '$d.backgrounds[$i|tonumber]')"
 
-			inkscape --without-gui \
-				-w "${size}" -h "${size}" \
-				-b "${background}" \
-				-e "${TARGET_DIR}/${name}_background_x${size}.png" \
-				"${SCRIPT_DIR}/${file}"
+			heightIndex='0'
+			while [ "${heightIndex:?}" -le "${heightTotal:?}" ]; do
+				height="$(jq -nr --argjson d "${vector:?}" --arg i "${heightIndex:?}" '$d.heights[$i|tonumber]')"
 
-			if [ "${margin}" != 'null' ]; then
-				inkscape --without-gui \
-					-w "${size}" -h "${size}" \
-					-a "${margin}" \
-					-e "${TARGET_DIR}/${name}_margin_x${size}.png" \
-					"${SCRIPT_DIR}/${file}"
+				marginIndex='0'
+				while [ "${marginIndex:?}" -le "${marginTotal:?}" ]; do
+					margin="$(jq -nr --argjson d "${vector:?}" --arg i "${marginIndex:?}" '$d.margins[$i|tonumber]')"
 
-				inkscape --without-gui \
-					-w "${size}" -h "${size}" \
-					-b "${background}" \
-					-a "${margin}" \
-					-e "${TARGET_DIR}/${name}_background_margin_x${size}.png" \
-					"${SCRIPT_DIR}/${file}"
-			fi
+					inkscapeOpts="--export-filename=${TARGET_DIR:?}/${name:?}-${background#'#'}-h${height:?}-(${margin:?}).png"
+					inkscapeOpts="${inkscapeOpts:?} --export-height=${height:?}"
+					inkscapeOpts="${inkscapeOpts:?} --export-area=${margin:?}"
+					if [ "${background:?}" != 'transparent' ]; then
+						inkscapeOpts="${inkscapeOpts:?} --export-background=${background:?}"
+					fi
+
+					# shellcheck disable=SC2086
+					inkscape ${inkscapeOpts:?} "${SCRIPT_DIR:?}/${src:?}"
+
+					marginIndex="$((marginIndex+1))"
+				done
+
+				heightIndex="$((heightIndex+1))"
+			done
+
+			backgroundIndex="$((backgroundIndex+1))"
 		done
+
+		vectorIndex="$((vectorIndex+1))"
 	done
 }
 
-main
+main "${@-}"
